requirements.txt
arrow
eth-brownie
tokenlists
